source/requirements.txt
# requirments without version numbers
tkinter
pillow
os